+++ zetaclient/chains/sui/observer/observer_test.go
@@ -8,6 +8,7 @@ import (
 
 	"cosmossdk.io/math"
 	"github.com/block-vision/sui-go-sdk/models"
+	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -113,8 +114,8 @@ func TestObserver(t *testing.T) {
 		ts.suiMock.On("QueryModuleEvents", mock.Anything, expectedQuery).Return(events, "", nil)
 
 		// Given 2 transaction blocks
-		ts.OnGetTx("TX_1_ok", "10000", false, nil)
-		ts.OnGetTx("TX_3_ok", "20000", false, nil)
+		ts.OnGetTx("TX_1_ok", "10000", true, false, nil)
+		ts.OnGetTx("TX_3_ok", "20000", true, false, nil)
 
 		// Given inbound votes catches so we can assert them later
 		ts.CatchInboundVotes()
@@ -200,7 +201,7 @@ func TestObserver(t *testing.T) {
 		ts.suiMock.On("QueryModuleEvents", mock.Anything, expectedQuery).Return(events, "", nil)
 
 		// Given transaction block
-		ts.OnGetTx("TX_restricted", "10000", false, nil)
+		ts.OnGetTx("TX_restricted", "10000", true, false, nil)
 
 		// Given inbound votes catches so we can assert them later
 		ts.CatchInboundVotes()
@@ -237,7 +238,7 @@ func TestObserver(t *testing.T) {
 		// Given underlying tx with event
 		evmAlice := sample.EthAddress()
 
-		ts.OnGetTx("TX_TRACKER_1", "15000", true, []models.SuiEventResponse{
+		ts.OnGetTx("TX_TRACKER_1", "15000", true, true, []models.SuiEventResponse{
 			ts.SampleEvent("TX_TRACKER_1", string(sui.DepositEvent), map[string]any{
 				"coin_type": string(sui.SUI),
 				"amount":    "1000",
@@ -281,13 +282,14 @@ func TestObserver(t *testing.T) {
 
 		ts.MockCCTXByNonce(cctx)
 
-		// Given outbound tracker
+		// Given outbound tracker containing two tx hashes, one of which is false
 		const digest = "0xSuiTxHash"
+		const digest2 = "0xSuiTxHashFalse"
 		tracker := cctypes.OutboundTracker{
 			Index:    "0xAAA",
 			ChainId:  ts.Chain().ChainId,
 			Nonce:    nonce,
-			HashList: []*cctypes.TxHash{{TxHash: digest}},
+			HashList: []*cctypes.TxHash{{TxHash: digest2}, {TxHash: digest}},
 		}
 
 		ts.MockOutboundTrackers([]cctypes.OutboundTracker{tracker})
@@ -339,8 +341,10 @@ func TestObserver(t *testing.T) {
 				},
 			},
 		}
+		tx2 := models.SuiTransactionBlockResponse{Digest: digest2}
 
-		ts.MockGetTxOnce(tx)
+		ts.MockGetTxOnce(tx, nil)
+		ts.MockGetTxOnce(tx2, errors.New("no tx found"))
 
 		// ACT
 		err = ts.ProcessOutboundTrackers(ts.ctx)
@@ -591,14 +595,20 @@ func (ts *testSuite) SampleEvent(txHash, event string, kv map[string]any) models
 	}
 }
 
-func (ts *testSuite) OnGetTx(digest, checkpoint string, showEvents bool, events []models.SuiEventResponse) {
+func (ts *testSuite) OnGetTx(digest, checkpoint string, showEffects, showEvents bool, events []models.SuiEventResponse) {
 	req := models.SuiGetTransactionBlockRequest{
-		Digest:  digest,
-		Options: models.SuiTransactionBlockOptions{ShowEvents: showEvents},
+		Digest: digest,
+		Options: models.SuiTransactionBlockOptions{
+			ShowEffects: showEffects,
+			ShowEvents:  showEvents,
+		},
 	}
 
 	res := models.SuiTransactionBlockResponse{
-		Digest:     digest,
+		Digest: digest,
+		Effects: models.SuiEffects{
+			Status: models.ExecutionStatus{Status: client.TxStatusSuccess},
+		},
 		Events:     events,
 		Checkpoint: checkpoint,
 	}
@@ -606,8 +616,21 @@ func (ts *testSuite) OnGetTx(digest, checkpoint string, showEvents bool, events 
 	ts.suiMock.On("SuiGetTransactionBlock", mock.Anything, req).Return(res, nil).Once()
 }
 
-func (ts *testSuite) MockGetTxOnce(tx models.SuiTransactionBlockResponse) {
-	ts.suiMock.On("SuiGetTransactionBlock", mock.Anything, mock.Anything).Return(tx, nil).Once()
+func (ts *testSuite) MockGetTxOnce(tx models.SuiTransactionBlockResponse, err error) {
+	req := models.SuiGetTransactionBlockRequest{
+		Digest: tx.Digest,
+		Options: models.SuiTransactionBlockOptions{
+			ShowEvents:  true,
+			ShowInput:   true,
+			ShowEffects: true,
+		},
+	}
+
+	if err == nil {
+		ts.suiMock.On("SuiGetTransactionBlock", mock.Anything, req).Return(tx, err).Once()
+	} else {
+		ts.suiMock.On("SuiGetTransactionBlock", mock.Anything, req).Return(models.SuiTransactionBlockResponse{}, err).Once()
+	}
 }
 
 func (ts *testSuite) CatchInboundVotes() {