+++ zetaclient/chains/sui/signer/signer_tracker.go
@@ -70,6 +70,7 @@ func (s *Signer) reportOutboundTracker(ctx context.Context, nonce uint64, digest
 			// failed outbound should be ignored as it cannot increment the gateway nonce.
 			// Sui transaction status is one of ["success", "failure"]
 			// see: https://github.com/MystenLabs/sui/blob/615516edb0ed55e45d599f042f9570b493ce9643/crates/sui-json-rpc-types/src/sui_transaction.rs#L1345
+			logger.Error().Msgf("tx failed with error: %s", res.Effects.Status.Error)
 			return errors.Errorf("tx failed with error: %s", res.Effects.Status.Error)
 		case res.Effects.Status.Status == client.TxStatusSuccess && res.Checkpoint != "":
 			return s.postTrackerVote(ctx, nonce, digest)
@@ -83,6 +84,6 @@ func (s *Signer) reportOutboundTracker(ctx context.Context, nonce uint64, digest
 
 // note that at this point we don't care whether tx was successful or not.
 func (s *Signer) postTrackerVote(ctx context.Context, nonce uint64, digest string) error {
-	_, err := s.zetacoreClient.PostOutboundTracker(ctx, s.Chain().ChainId, nonce, digest)
+	_, err := s.zetaRepo.PostOutboundTracker(ctx, s.Logger().Std, nonce, digest)
 	return err
 }