+++ zetaclient/testutils/mocks/gen/sui.go
@@ -13,7 +13,7 @@ import (
 // client represents interface version of Client.
 // It's unexported on purpose ONLY for mock generation.
 //
-//go:generate mockery --name suiClient --structname SuiClient --filename sui_client.go --output ./
+//go:generate mockery --name suiClient --structname SuiClient --filename sui_client.go --output ../
 //nolint:unused // used for code gen
 type suiClient interface {
 	HealthCheck(ctx context.Context) (time.Time, error)
@@ -24,6 +24,10 @@ type suiClient interface {
 	GetSuiCoinObjectRefs(ctx context.Context, owner string, minBalanceMist uint64) ([]*suiptb.ObjectRef, error)
 
 	SuiXGetLatestSuiSystemState(ctx context.Context) (models.SuiSystemStateSummary, error)
+	SuiXGetDynamicFieldObject(
+		ctx context.Context,
+		req models.SuiXGetDynamicFieldObjectRequest,
+	) (models.SuiObjectResponse, error)
 	SuiXGetReferenceGasPrice(ctx context.Context) (uint64, error)
 	SuiXQueryEvents(ctx context.Context, req models.SuiXQueryEventsRequest) (models.PaginatedEventsResponse, error)
 	SuiMultiGetObjects(ctx context.Context, req models.SuiMultiGetObjectsRequest) ([]*models.SuiObjectResponse, error)